requirements.txt
flask
flask-wtf
requests
python-nmap
python-dotenv
pandas
fpdf